benchtests: Add pthread adaptive spin mutex microbenchmark
Add a microbenchmark for measuring mutex lock and unlock performance with
varying numbers of threads and varying size of a critical section. The
benchmark leverages the mutex lock and unlock operation for protecting the
critical section and measures the minimum iterations, the maximum
iterations, and the total iterations within a fixed duration. Variants of
benchmark are run with 1, 2, 3, 4, nproc/4, nproc/2, nproc threads.

The size of critical section is determined by the number of times of pause
(x86 only) and read which is intended to emulate the scenarios in real
applications. In this microbenchmark, the number 1, 10, 100, and 1000 are
used to represent different size of critical sections in the working set.

    * benchtests/bench-mutex-adaptive-thread.c: Microbenchmark for
    adaptive spin mutex
    * benchmark/Makefile: Add adaptive spin mutex benchmark

ChangLog:
    V2->V3:
    a) Add some delay after mutex unlock to reduce the possibility of lock
    acquisition by the previous lock holder, and that makes more sense
    for practical applications
    b) Sync threads to start simultaneously
    c) Set CPU affinity for threads

    V1->V2:
    New added microbenchmark, as requested by Adhemerval Zanella

Signed-off-by: Kemi Wang <kemi.wang@intel.com>