add mcs_lock.h and mcs_lock.c
The overhead of lock acquisition increases linearly with the contending
thread number due to

The  Since the
cache line including the lock is shared among all of the spinners, when
lock is released, each spinner will try to acquire the lock via cmpxchg
instruction which constantly floods the system via "read-for-ownership"
requests. As a result, there will be a lot of cache line bouncing in a
large system with a lots of CPUs.

Signed-off-by: Kemi Wang <kemi.wang@intel.com>