add mcs_lock.h and mcs_lock.c
Use MCS-style lock to queue spinner, thus, the overhead of lock holder
transformation keeps O(1) (the cache line of the lock is shared between
current lock holder and the spinner which is spinning on spinlock).

However, lock holder preemption issue would potentially cause lock
performance collapse when CPUs are oversubscribed. Thus, we introduce a
timeout tunable for each spinner spinning on local flag in the queue. If a
spinner has spinning on the local flag for exceeding a specified level of
time in the queue, it is allowed to spin on the spin lock even if the flag
has not been set by its predecessor.

Signed-off-by: Kemi Wang <kemi.wang@intel.com>