lock: introduce queue lock mechanism
Three monthes ago, I proposed a queue spinners algorithm based-on legacy mcs
style lock. To avoid lock performance collapse when CPUs are oversubscribed,
this algorithm allows a long waited spinner leave the queue if timeout by
introducing a threshold of spin counts. According to the feedback from
internally and customers, one main disadvantage of this algorithm is there
is an extra platform-specific tunables which leaves many tuning work to
system adminstrators and makes code hard to maintain. Therefore, I revisted
this issue and proposed a new one as below.

==========================Cut Here=======================================

The poor performance of simple busy-waiting spinlock by spinning on read
(test and test-and-set) has been researched in many literatures[1,2,3].
(The way for busy-waiting spinlock via spinning on test-and-set is even
worse).

Note: Depended on architecture-specific implementation, this test-and-set
can be cmpxchg or other atomic instruction. But, all of them suffers
similar issues.

Todo: spinlock performance decreases with the increase of contending thread
number, we will submit a bug at redhat bugzilla.

The main problem behind current spinlock implementation can be summarized
as below:
1) when the lock becomes available, all the spinnning tasks will try to
acquire this spinlock more or less simultaneously. There will be a lot of
cache line bouncing especially on systems with a large number of CPUs;

2) Each atomic operation will trigger a read-for-ownership request which
invalidates all of the cache copies in CPUs before executing the atomic
operation semantics. In an invalidation-based cache coherence system,
depended on cache coherency hardware design, it requires O(*N*) time or the
largest time for invalidating cache line in other core to broadcast cache
line invalidation to *N* waiting tasks.

3) When the lock becomes available, these cache copies will be invalidated,
each processor will then incur a read miss to fetch the new value back into
its cache. These read misses will be satisfied serially. Each process to
get the new value will then try to execuate a atomic test-and-set. These
requests must compete for the bus resource with other processors doing read
miss.

To address these above problems, MCS style lock [3] as a well-known
solution is proposed for that by queue spinners. However, while processor
preemption can lead to bad lock performance [4], this queue spinner way
makes it much worse (we have seen lock performance collapse when CPUs are
oversubscribed). For instance, when current spinner has acquired the lock
and wake up the next spinner to spin on the lock, but it is not running
(the CPU is scheduled to run other task). Thus, the other spinners have to
wait in the queue until that spinner is rechedualed back on the CPU.

Therefore, we make some minor change based-on legacy MCS style lock to
avoid this problem. Compare to the legacy way, the main difference is that
we change the way for waking up spinner from only wake up the next one to
wake up one or multiple spinners serially until we saw a spinner has
been spinning on the lock. To achieve this target, we introduce a global
token to identify the activity of a spinner, this token is set to 0 as
default at the beginning, if a spinner want to spin on the lock, it must
get that token first. For instance, when a spinner wakes up the next
spinner, the former does not know its following is running or not, if the
next spinner is running, it will get that token immediately by setting to 1
via cmpxchg instruction, if not, this token is still 0, and the former will
continue to wake up the remianing spinners until some one gets that token.

Test result:
On a 2-sockets Skylake platform using the tst-spin test which I posted
several days ago. Compare to the existed spinlock,
a) this patch performs slight performance degradation about 6% in less
contention scenario, this is expected because our code introduces extra
code complexity and the overhead of lock holder transformation is not so
high in that case.
b) In high contention scenario, this patch can reach up to about 3x performance
improvement.
c) In the case of CPU oversubscribed with high lock contention, the
scenarios is emulated by running 2 processes in parallel, each process has
56 threads running with setting CPU affinity accordingly. we does not
see lock performance collapse, instead, the throughput improves about
2x.

Todo:
a) will submit this spinlock test case to upstream
b) will post more data of test result

Referrence:
[1] Anderson, Thomas E. "The performance of spin lock alternatives for
shared-money multiprocessors." IEEE Transactions on Parallel and
Distributed Systems 1.1 (1990): 6-16.

[2] Boyd-Wickizer, Silas, et al. "Non-scalable locks are dangerous."
Proceedings of the Linux Symposium. 2012.

[3] Mellor-Crummey, John M., and Michael L. Scott. "Algorithms for scalable
synchronization on shared-memory multiprocessors." ACM Transactions on
Computer Systems (TOCS) 9.1 (1991): 21-65.

[4] Zahorjan, John. "Spinning versus blocking in parallel systems with
uncertainty." Proceedings of the International Seminar on Performance of
Distributed and Parallel Systems. 1988.

Signed-off-by: Kemi Wang <kemi.wang@intel.com>