Enlarge MAX_ADAPTIVE_SPIN to 10000
Though, it's hard to determine the optimal value of this timeout tunable
interface to acquire the best system performance. We can default the value
to fit most cases at least.

Simply, this tunable should be greater than the normal waiting time, and
the maximum of which can be calculated by contending thread number * the
time consumed in a cirtical section size. Image we have 1000 threads
contending on the lock at the same time, and the average time consumed in
the critical section is 1000ns. The maximum waiting time in the queue (the
1000st spinner) is 1000*1000ns=1ms. IOW, if a spinner in the queue has
waited for exceeding 1ms, the preemption probably happens, and it's
unnecessary for spinners to go on spinning on the local flag in the queue.

In our patch, the timeout is measured by the number of executing pause
instruction. We know 1 pause instruction is about 140 cycles(~45ns) on
Skylake platform. Thus, the number of spin count is about 20000.

Depend on specific size of critical section and thread number, this value
can be tuned to meet the optimal system performance.

Todo: Add a tunable interface for that

Signed-off-by: Kemi Wang <kemi.wang@intel.com>